# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingCreate": {
        "properties": {
          "check_in": {
            "format": "date",
            "title": "Check In",
            "type": "string"
          },
          "check_out": {
            "format": "date",
            "title": "Check Out",
            "type": "string"
          },
          "guests": {
            "default": 1,
            "minimum": 1.0,
            "title": "Guests",
            "type": "integer"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          }
        },
        "required": [
          "listing_id",
          "check_in",
          "check_out"
        ],
        "title": "BookingCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentCreate": {
        "properties": {
          "booking_id": {
            "title": "Booking Id",
            "type": "string"
          },
          "kind": {
            "default": "other",
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "booking_id"
        ],
        "title": "IncidentCreate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Managed hospitality platform. This spec is generated from the implementation (code-first) and kept in sync by a CI drift-guard — see docs/api/README.md.",
    "title": "Homies API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/audit": {
      "get": {
        "operationId": "list_audit_v1_admin_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/bookings": {
      "get": {
        "operationId": "list_bookings_v1_admin_bookings_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Bookings",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/founder-feed": {
      "get": {
        "description": "Founder operational visibility: in-app notifications routed to the\nfounder role, newest first.",
        "operationId": "founder_feed_v1_admin_founder_feed_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Founder Feed",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/incidents": {
      "get": {
        "operationId": "list_incidents_v1_admin_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Incidents",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "open_incident_v1_admin_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open Incident",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/kpi": {
      "get": {
        "description": "Monetary KPIs, answered from the ledger (OBS-07).\n\nThe window is half-open [from, to) so consecutive periods never\ndouble-count a boundary day. Money is in integer minor units and ratios in\nbasis points — see app/modules/admin/kpi.py for why neither is a float.",
        "operationId": "kpi_v1_admin_kpi_get",
        "parameters": [
          {
            "description": "Inclusive start (UTC date)",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "description": "Inclusive start (UTC date)",
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "description": "EXCLUSIVE end (UTC date)",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "description": "EXCLUSIVE end (UTC date)",
              "format": "date",
              "title": "To",
              "type": "string"
            }
          },
          {
            "description": "Defaults to the platform currency",
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to the platform currency",
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Kpi",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ledger/balances": {
      "get": {
        "operationId": "ledger_balances_v1_admin_ledger_balances_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ledger Balances",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ledger/entries": {
      "get": {
        "operationId": "list_ledger_entries_v1_admin_ledger_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Ledger Entries",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ledger/reconciliation": {
      "get": {
        "operationId": "ledger_reconciliation_v1_admin_ledger_reconciliation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ledger Reconciliation",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/notifications": {
      "get": {
        "description": "Founder delivery audit (OAT-03): status, retry count, channel, last\nerror, timestamps. Filter status=dead (or failed) for the dead-letter view.",
        "operationId": "list_notifications_v1_admin_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/notifications/queue": {
      "get": {
        "description": "Queue depth by delivery state — founder observability without SQL.",
        "operationId": "notification_queue_v1_admin_notifications_queue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Notification Queue",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/payments": {
      "get": {
        "operationId": "list_payments_v1_admin_payments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Payments",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/payments/reconciliation": {
      "get": {
        "description": "Payment<->ledger consistency + (when Stripe keys are present) a\nStripe-balance cross-check. Used as the daily reconciliation report.",
        "operationId": "payments_reconciliation_v1_admin_payments_reconciliation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Payments Reconciliation",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/properties/{property_id}/authorities": {
      "get": {
        "operationId": "property_authorities_v1_admin_properties__property_id__authorities_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Property Authorities",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/property-authorities/{authority_id}/revoke": {
      "post": {
        "operationId": "revoke_property_authority_v1_admin_property_authorities__authority_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "authority_id",
            "required": true,
            "schema": {
              "title": "Authority Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Property Authority",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/property-authorities/{authority_id}/verify": {
      "post": {
        "operationId": "verify_property_authority_v1_admin_property_authorities__authority_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "authority_id",
            "required": true,
            "schema": {
              "title": "Authority Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Property Authority",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "list_users_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/bookings": {
      "get": {
        "operationId": "my_bookings_v1_bookings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "title": "Response My Bookings V1 Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Bookings",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "create_booking_v1_bookings_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
  },
...
}